api/requirements.txt
# FastAPI Core
fastapi>=0.127.0
uvicorn[standard]==0.40.0
python-multipart

# Database
sqlmodel
asyncpg
psycopg2-binary

# HTTP Client
httpx==0.28.1

# Web Automation & Testing
playwright
pytest-playwright
# Utilities
python-dotenv
pydantic==2.12.5
pydantic-settings
email-validator
jinja2>=3.1.0  # Templates para prompts

# Security — Keycloak JWT (RS256)
python-jose[cryptography]

##seo
advertools
pandas  # advertools usa pandas internamente
trafilatura
beautifulsoup4
lxml    # parser rápido para bs4

reportlab
openpyxl
playwright-stealth
nodriver
pyvirtualdisplay

sqlalchemy

# Monitoring
prometheus-client
prometheus-fastapi-instrumentator

# Office Documents
pdf2docx

# Schema Extraction
extruct

# Schema Validation (JSON-LD + Schema.org)
pyld>=2.0.4
# Nota: pydantic-schemaorg 1.x requiere pydantic<2 y rompe FastAPI moderno.
# Si se usa, debe instalarse en un entorno separado o con stack legacy.

# Tokenizer
tiktoken


